Fixed #302
added an external method that determines if an authenticated user is a membrane user or a standard acl_users user, and sends them to the appropriate location for 'My Folder'.

Added GS import steps to register this method, and to overwrite the standard 'My Folder' action in plone.  

=== To Be Removed SOON ===
Temporary modification to Person allows me to drop into pdb in my terminal when I click on the vcard export button (just commented out most of that function and replaced it with a pdb.set_trace().