Okay, fixed the problem for real this time.
Instead of just hiding the stupid, useless relations tab, let's actually change the permissions needed to see it.  We'll set it up so that it's only visible to people who are actually supposed to be able to set up the relations we are trying to control, that is to say, managers and personnel managers.  So I will use the perms we use for adding the relations in the first place.

As a side note, it would be really nifty if we could set the value of 'originalRelationsPerms' in AppInstall.py from some config value in the original relations product.  This fix is potentially brittle in that if that permission is changed, our fix will return the action to an incorrect state, but there's nothing we can do about that right now.

fixes #400 (again)